+++ redisson/src/main/java/org/redisson/api/RMapCache.java
@@ -52,7 +52,7 @@ public interface RMapCache<K, V> extends RMap<K, V>, RMapCacheAsync<K, V> {
      * @param maxSize - max size
      *                  If <code>0</code> the cache is unbounded (default).
      */
-    void setMaxSize(int maxSize);
+    void setMaxSize(long maxSize);
 
     /**
      * Sets max size of the map and overrides current value.
@@ -61,7 +61,7 @@ public interface RMapCache<K, V> extends RMap<K, V>, RMapCacheAsync<K, V> {
      * @param maxSize - max size
      * @param mode - eviction mode
      */
-    void setMaxSize(int maxSize, EvictionMode mode);
+    void setMaxSize(long maxSize, EvictionMode mode);
 
     /**
      * Tries to set max size of the map. 
@@ -71,7 +71,7 @@ public interface RMapCache<K, V> extends RMap<K, V>, RMapCacheAsync<K, V> {
      * @return <code>true</code> if max size has been successfully set, otherwise <code>false</code>.
      *         If <code>0</code> the cache is unbounded (default).
      */
-    boolean trySetMaxSize(int maxSize);
+    boolean trySetMaxSize(long maxSize);
 
     /**
      * Tries to set max size of the map.
@@ -81,7 +81,7 @@ public interface RMapCache<K, V> extends RMap<K, V>, RMapCacheAsync<K, V> {
      * @param mode - eviction mode
      * @return <code>true</code> if max size has been successfully set, otherwise <code>false</code>.
      */
-    boolean trySetMaxSize(int maxSize, EvictionMode mode);
+    boolean trySetMaxSize(long maxSize, EvictionMode mode);
     /**
      * If the specified key is not already associated
      * with a value, attempts to compute its value using the given mapping function and enters it into this map .
